JSON: changed child(uinteger) function to return JSONpair so key and value can be retrieved. File: added new Class ZipFile to manage zip archives.
At the moment the ZipFile class is only able to read files < 2gb from an existing zip archive. Eventually it will be a full fledged Zip archive manager with the functions one would expect. This makes use of the recent MemoryFileDriver addition to the File class. Files are loaded into memory so the file should be closed when no longer required to reclaim the memory.