Added support for non-x86 builds
Builds for non-x86 architectures should now be possible.
The architecture of the target is automatically determined by
the default fbc target. Currently support is limited to
arm architectures. This value can be specified at build
time by passing the ARCH variable with the value of x86 or arm.

Example:
make ARCH=x86

Currently this only set the compiler option for inline asm syntax.